requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
boto3==1.34.32
botocore==1.34.32
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
config==0.5.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
fire==0.7.0
fonttools==4.54.1
future==0.18.3
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jmespath==1.0.1
kombu==5.3.5
lxml==5.3.0
MarkupSafe==2.1.4
numpy==2.1.2
openapi-codec==1.3.2
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==23.2
pandas==2.2.3
pdf2docx==0.5.8
phonenumbers==8.13.29
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
PyMuPDF==1.24.11
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2023.4
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
termcolor==2.5.0
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
